% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{category_contribution}
\alias{category_contribution}
\title{Plot the contribution of different categories to a measure}
\usage{
category_contribution(data, category, measure)
}
\arguments{
\item{data}{A data frame or tibble}

\item{category}{Unquoted name of category (can be factor, character or numeric)}

\item{measure}{Unquoted name of measure}
}
\value{
A ggplot plot object
}
\description{
Plot the contribution of different categories to a measure
}
\examples{
category_contribution(ggplot2::diamonds, cut, price)
category_contribution(ggplot2::diamonds, clarity, price)
}
