% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_to_factor}
\alias{col_to_factor}
\title{Private utility function: given a possibly non-factor column passed as a quosure, convert into a factor}
\usage{
col_to_factor(data, col_enquo)
}
\arguments{
\item{data}{A data frame or tibble}

\item{col_enquo}{A quosure}
}
\value{
A data frame or tibble with the corresponding column converted to factor if nevessary
}
\description{
Private utility function: given a possibly non-factor column passed as a quosure, convert into a factor
}
