% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_trends.R
\name{measure_change_over_time_wide}
\alias{measure_change_over_time_wide}
\title{Plot the change of a measure (or set of measures) over time where each measure is in a different column}
\usage{
measure_change_over_time_wide(data, time_col, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{time_col}{Unquoted column name with time values to plot on the x axis}

\item{...}{Unquoted column names of one or more measures to plot (up to 6 measures)}
}
\value{
A ggplot plot object
}
\description{
Plot the change of a measure (or set of measures) over time where each measure is in a different column
}
\examples{
measure_change_over_time_wide(ggplot2::economics, date, pop, unemploy)
}
