% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-distribution.R
\name{measure_distribution_by_two_categories}
\alias{measure_distribution_by_two_categories}
\title{Plot the distribution of a numeric (measure) column differentiated by two categories}
\usage{
measure_distribution_by_two_categories(
  data,
  measure,
  category1,
  category2,
  bwidth = NULL
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{measure}{Unquoted column name of containing numbers (measure)}

\item{category1, category2}{Unquoted column names of categories (can be factor, character or numeric)}

\item{bwidth}{width of bin for histogram (by default uses binwidth for 30 bins)}
}
\value{
A ggplot plot object
}
\description{
Plot the distribution of a numeric (measure) column differentiated by two categories
}
\examples{
measure_distribution_by_two_categories(ggplot2::mpg, hwy, class, fl)
measure_distribution_by_two_categories(ggplot2::diamonds, carat, cut, clarity)
}
