% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-relationships.R
\name{multi_measures_relationship}
\alias{multi_measures_relationship}
\title{Plot the relationship between many measures}
\usage{
multi_measures_relationship(data, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{Unquoted column names of numeric columns (measures)}
}
\value{
A ggplot plot object
}
\description{
Plot the relationship between many measures
}
\examples{
multi_measures_relationship(ggplot2::mpg, hwy, displ)
multi_measures_relationship(ggplot2::mpg, cty, hwy, displ)
}
