% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallies.R
\name{category_tally}
\alias{category_tally}
\title{Plot counts of a category}
\usage{
category_tally(data, category_column)
}
\arguments{
\item{data}{A data frame or tibble}

\item{category_column}{Unquoted column name of category (can be factor, character or numeric)}
}
\value{
A ggplot plot object
}
\description{
Plot counts of a category
}
\examples{
category_tally(ggplot2::mpg, class)
category_tally(ggplot2::diamonds, cut)
}
